Fix parsing the Arbiter brick type (#10)
Arbiter brick type was identified as regular brick. Due to this
the `subvol.total_size was min(bricks total sizes)`. subvol free
size was calculated as `total - used`, but this value was negative
because of `total` is less than `used`.

This issue is seen only when Arbiter brick size was not equal to
other bricks (Test setup).

Signed-off-by: Aravinda Vishwanathapura <mail@aravindavk.in>